Recuperacion Proyecto

+++ src/controlador/PedidoJpaController.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package controlador;
 
 import controlador.exceptions.IllegalOrphanException;
@@ -18,24 +13,14 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
 import modelo.Pedido;
 
-/**
- *
- * @author RICARDO
- */
 public class PedidoJpaController implements Serializable {
 
-    private EntityManagerFactory emf;
-    
     public PedidoJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-
-    public PedidoJpaController() {
-        this.emf = Persistence.createEntityManagerFactory("SistemaHospitalarioPU");
-    }
+    private EntityManagerFactory emf = null;
 
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
@@ -116,6 +101,14 @@ public class PedidoJpaController implements Serializable {
                     illegalOrphanMessages.add("The Consulta " + consultaNew + " already has an item of type Pedido whose consulta column cannot be null. Please make another selection for the consulta field.");
                 }
             }
+            for (Examen listaExamenOldExamen : listaExamenOld) {
+                if (!listaExamenNew.contains(listaExamenOldExamen)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Examen " + listaExamenOldExamen + " since its pedido field is not nullable.");
+                }
+            }
             if (illegalOrphanMessages != null) {
                 throw new IllegalOrphanException(illegalOrphanMessages);
             }
@@ -138,12 +131,6 @@ public class PedidoJpaController implements Serializable {
             if (consultaNew != null && !consultaNew.equals(consultaOld)) {
                 consultaNew.setPedido(pedido);
                 consultaNew = em.merge(consultaNew);
-            }
-            for (Examen listaExamenOldExamen : listaExamenOld) {
-                if (!listaExamenNew.contains(listaExamenOldExamen)) {
-                    listaExamenOldExamen.setPedido(null);
-                    listaExamenOldExamen = em.merge(listaExamenOldExamen);
-                }
             }
             for (Examen listaExamenNewExamen : listaExamenNew) {
                 if (!listaExamenOld.contains(listaExamenNewExamen)) {
@@ -173,7 +160,7 @@ public class PedidoJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(Long id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -185,15 +172,21 @@ public class PedidoJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The pedido with id " + id + " no longer exists.", enfe);
             }
+            List<String> illegalOrphanMessages = null;
+            List<Examen> listaExamenOrphanCheck = pedido.getListaExamen();
+            for (Examen listaExamenOrphanCheckExamen : listaExamenOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Pedido (" + pedido + ") cannot be destroyed since the Examen " + listaExamenOrphanCheckExamen + " in its listaExamen field has a non-nullable pedido field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
             Consulta consulta = pedido.getConsulta();
             if (consulta != null) {
                 consulta.setPedido(null);
                 consulta = em.merge(consulta);
-            }
-            List<Examen> listaExamen = pedido.getListaExamen();
-            for (Examen listaExamenExamen : listaExamen) {
-                listaExamenExamen.setPedido(null);
-                listaExamenExamen = em.merge(listaExamenExamen);
             }
             em.remove(pedido);
             em.getTransaction().commit();
@@ -250,29 +243,4 @@ public class PedidoJpaController implements Serializable {
         }
     }
     
-    public List<Pedido> getPedidosPorEstado(String estado) {
-        EntityManager em = getEntityManager();
-        try {
-            Query q = em.createQuery("SELECT e "
-                    + "FROM Pedido e " 
-                    + "WHERE e.estado_pedido = ?1")
-                    .setParameter(1, estado);
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-    
-    public List<Pedido> getPedidosPorTodosMenosUnEstado(String estado) {
-        EntityManager em = getEntityManager();
-        try {
-            Query q = em.createQuery("SELECT e "
-                    + "FROM Pedido e " 
-                    + "WHERE NOT e.estado_pedido = ?1")
-                    .setParameter(1, estado);
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
 }
